Add level 8 features New classes Parser(interpret user command), Storage(create file and store data), TaskList(list of tasks) and ui(response to users) are implemented.
