+++ src/main/java/Task.java
@@ -1,10 +1,12 @@
 public class Task {
     protected String description;
     protected boolean isDone;
+    protected int sequence;
 
     public Task(String description) {
         this.description = description;
         this.isDone = false;
+        this.sequence = 0;
     }
 
     public String getStatusIcon() {
